Improve type consistency when counting contributors.
`int`, `uint32` and `uint64` were all mixed together.
This standardizes on `uint32`, since that is the max count
supported by the hash table.

+++ pg_diffix/template/contribution_state.h
@@ -113,9 +113,9 @@ typedef struct CS_TOP_CONTRIBUTOR
 
 typedef struct CS_CONTRIBUTION_STATE
 {
-  uint64 distinct_aids;            /* Number of distinct AIDs being tracked */
-  uint64 distinct_contributors;    /* Number of distinct AIDs that contributed at least once */
-  uint64 total_contributions;      /* Total number of contributions */
+  uint32 distinct_aids;            /* Number of distinct AIDs being tracked */
+  uint32 distinct_contributors;    /* Number of distinct AIDs that contributed at least once */
+  uint32 total_contributions;      /* Total number of contributions */
   uint32 aid_seed;                 /* Seed derived from unique AIDs */
   MemoryContext context;           /* Where the hash table lives */
   CS_TABLE_HASH *all_contributors; /* All contributors (map of AID -> TableEntry) */
@@ -125,7 +125,7 @@ typedef struct CS_CONTRIBUTION_STATE
 #ifdef CS_OVERALL_CONTRIBUTION_CALCULATE
   CS_CONTRIBUTION_TYPE overall_contribution; /* Total contribution from all contributors */
 #endif
-  int top_contributors_length;                                /* Length of top_contributors array */
+  uint32 top_contributors_length;                             /* Length of top_contributors array */
   CS_TOP_CONTRIBUTOR top_contributors[FLEXIBLE_ARRAY_MEMBER]; /* Stores top_contributors_length number of top contributors */
 
 #endif /* CS_TRACK_CONTRIBUTION */
@@ -137,7 +137,7 @@ CS_SCOPE CS_CONTRIBUTION_STATE *CS_STATE_NEW(
     MemoryContext context
 #ifdef CS_TRACK_CONTRIBUTION
     ,
-    int top_contributors_length
+    uint32 top_contributors_length
 #endif
 );
 
@@ -244,12 +244,12 @@ static inline int CS_AID_INDEX(
 
 static inline void CS_INSERT_CONTRIBUTOR(
     CS_CONTRIBUTION_STATE *state,
-    int top_length,
+    uint32 top_length,
     CS_AID_TYPE aid,
     CS_CONTRIBUTION_TYPE contribution)
 {
-  int insertion_index = CS_INSERTION_INDEX(state, top_length, contribution);
-  int capacity = state->top_contributors_length;
+  uint32 insertion_index = CS_INSERTION_INDEX(state, top_length, contribution);
+  uint32 capacity = state->top_contributors_length;
   size_t elements;
 
   if (insertion_index == capacity)
@@ -274,13 +274,13 @@ static inline void CS_INSERT_CONTRIBUTOR(
 
 static inline void CS_BUMP_OR_INSERT_CONTRIBUTOR(
     CS_CONTRIBUTION_STATE *state,
-    int top_length,
+    uint32 top_length,
     CS_AID_TYPE aid,
     CS_CONTRIBUTION_TYPE old_contribution,
     CS_CONTRIBUTION_TYPE new_contribution)
 {
-  int aid_index = CS_AID_INDEX(state, top_length, aid, old_contribution);
-  int insertion_index;
+  uint32 aid_index = CS_AID_INDEX(state, top_length, aid, old_contribution);
+  uint32 insertion_index;
   size_t elements;
 
   if (aid_index == top_length)
@@ -313,7 +313,7 @@ CS_SCOPE CS_CONTRIBUTION_STATE *CS_STATE_NEW(
     MemoryContext context
 #ifdef CS_TRACK_CONTRIBUTION
     ,
-    int top_contributors_length
+    uint32 top_contributors_length
 #endif
 )
 {
@@ -398,7 +398,7 @@ CS_SCOPE void CS_STATE_UPDATE_CONTRIBUTION(
 {
   bool found;
   uint32 aid_hash;
-  int top_length;
+  uint32 top_length;
   CS_TABLE_ENTRY *entry;
   CS_CONTRIBUTION_TYPE contribution_old;
   CS_CONTRIBUTION_TYPE min_top_contribution;
